requirements.txt
Django==5.1.4
openai==1.66.2
PyMuPDF==1.22.5
azureopenai==0.0.1
asgiref==3.8.1
sqlparse==0.5.3
Pillow==10.0.0
dotenv==0.9.9
pdf2image==1.16.3
azure-storage-blob==12.25.0

# azure-cognitiveservices-vision-computervision==0.9.1
# azure-ai-ml==1.25.0
# azure-ai-formrecognizer==3.2.0
# azure-ai-documentintelligence==1.0.0
gunicorn==20.1.0